Split execute into own file
  README: Add usage as module
  tests: test_main tests execute.py
